+++ tests/XmlTest.php
@@ -2,7 +2,7 @@
 /**
  * @copyright Zicht online <http://zicht.nl>
  */
-namespace ZichtTest\Util;
+namespace Zicht\Util\Test;
 
 use PHPUnit\Framework\TestCase;
 use Zicht\Util\Xml;